+++ src/euphorie/client/browser/settings.py
@@ -140,7 +140,12 @@ class Preferences(AutoExtensibleForm, form.Form):
     schema = PreferencesSchema
 
     show_personal_details = True
-    show_notifications = False
+
+    @property
+    def show_notifications(self):
+        return api.portal.get_registry_record(
+            "euphorie.notifications__enabled", default=False
+        )
 
     @property
     @memoize