+++ src/euphorie/client/notifications/send_notifications.py
@@ -1,5 +1,6 @@
 from euphorie.client.interfaces import INotificationCategory
 from euphorie.client.interfaces import INTERVAL_DAILY
+from plone import api
 from plone.protect.interfaces import IDisableCSRFProtection
 from Products.Five import BrowserView
 from zope.component import getAdapters
@@ -8,6 +9,10 @@ from zope.interface import alsoProvides
 
 class SendNotificationsDaily(BrowserView):
     def do_send(self):
+        if api.portal.get_registry_record(
+            "euphorie.notifications__enabled", default=False
+        ):
+            return
         notifications = getAdapters((self.context, self.request), INotificationCategory)
         for notification in notifications:
             if not notification[1].available: