+++ docs/changes.rst
@@ -5,6 +5,8 @@ Changelog
 -------------------
 
 - Notifications:
+  - Allow to enable/disable notifications via registry settings.
+    Note: the notification system is per default disabled.
   - Allow to customize the default's subscription setting for "ra not modified".
   - Allow to customize the notification mail's sender name and address.
   - Add translations.
@@ -13,7 +15,6 @@ Changelog
   - Add script to send notifications via cron.
   - Add basic notifications infrastructure.
   - Allow to set notification subscription settings in the preferences panel.
-    Note: the notification subscriptions are per default hidden.
   - Add BaseEmail class to unify sending emails.
   Ref: scrum-1572
   [thet, reinhardt]